Delete old version and renamed file
